fix: fix handling of database access failure not to hang when storages are inaccessible

+++ src/layer/infrastructure/indexeddb/model/transition.ts
@@ -1,6 +1,7 @@
 import { indexedDB } from '../module/global';
 import { states, commands, Command, InitialState, BlockState, UpgradeState, SuccessState, ErrorState, AbortState, CrashState, DestroyState, EndState } from './state';
 import { idbEventStream_, IDBEvent, IDBEventType } from './event';
+import { verifyStorageAccess } from '../../environment/api';
 import { causeAsyncException } from 'spica/exception';
 
 export function handle(database: string): void {
@@ -171,6 +172,7 @@ function handleFromCrashState(state: CrashState): void {
 
 function handleFromDestroyState(state: DestroyState): void {
   if (!state.alive) return;
+  if (!verifyStorageAccess()) return void handleFromEndState(new EndState(state));
   const { database } = state;
   const deleteRequest = indexedDB.deleteDatabase(database);
   deleteRequest.onsuccess = () => (
@@ -187,6 +189,7 @@ function handleFromEndState(state: EndState): void {
   void state.complete();
   void idbEventStream_
     .emit([database, IDBEventType.disconnect], new IDBEvent(database, IDBEventType.disconnect));
+  if (!verifyStorageAccess()) return;
   switch (commands.get(database) || command) {
     case Command.open:
       return void handleFromInitialState(new InitialState(database, version));
